pyproject.toml
[project]
name = "django-gan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "django-compressor>=4.5.1",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "requests==2.27.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.48.3",
]
